[BOLT][PE/COFF] Fix bounds checks, I/O error handling, and state rollback
Bug fixes:

1. PECOFFRewriteInstance: VAToFileOffset now tracks SizeOfRawData separately
   from VirtualSize to avoid returning file offsets past section raw data
   when VirtualSize > SizeOfRawData (legal in PE).

2. PECOFFRewriteInstance: Add missing bounds check (*OrigFileOff + PrologSize
   <= FileData.size()) before memcmp in prolog verification, matching the
   guard already present in freezePrologInstructions().

3. PDBRewriter: Cast loop index to uint64_t before multiplication to prevent
   uint32_t overflow on corrupt PDB data with large NumLines values.

4. LLD COFF LTO: Use |= instead of = for BBAddrMap so that
   /lto-basic-block-sections:labels is not silently overwritten.

5. DataReader::writeBranchProfile: Check OutFile.has_error() before
   returning success to detect disk-full or I/O errors.

6. InferEdgeCounts: Save and restore original block counts when verify()
   fails, preventing downstream passes from consuming half-propagated
   inconsistent profile data.

7. BinaryContext::getDataSection: Add explicit MachO branch instead of
   falling through to ELF path, matching getCodeSection pattern.

+++ bolt/lib/Rewrite/PDBRewriter.cpp
@@ -299,7 +299,7 @@ void PDBRewriter::rewritePDB(StringRef InputExe, StringRef OutputExe,
         uint64_t LineStart = FileBlockPos + 12;
 
         for (uint32_t L = 0; L < NumLines; ++L) {
-          uint64_t EntryPos = LineStart + L * 8;
+          uint64_t EntryPos = LineStart + (uint64_t)L * 8;
           if (EntryPos + 8 > DataEnd)
             break;
